src/data/preprocessing.py
import yfinance as yf
import pandas as pd
import numpy as np
from sklearn.preprocessing import StandardScaler
from sklearn.covariance import LedoitWolf
from sklearn.feature_selection import mutual_info_regression
import torch
from torch_geometric.data import Data

def download_price_data(tickers, start_date, end_date, price_data_file):
    """Download historical price data or load from pickle if available."""
    # Implementation here...

def calculate_returns(price_df):
    """Calculate daily returns from price data."""
    return price_df.pct_change().dropna()

def generate_features(returns_df, window_sizes=[5, 10, 20]):
    """Generate enhanced features for the returns data."""
    # Implementation here...

def create_sophisticated_edge_index(returns, correlation_threshold):
    """Create edge index based on partial correlations and mutual information."""
    # Implementation here... 