requirements.txt
torch>=1.9.0
torch-geometric>=2.0.0
pandas>=1.3.0
numpy>=1.19.0
scikit-learn>=0.24.0
yfinance>=0.1.63
matplotlib>=3.4.0
pyyaml>=5.4.0 